perf(maintenance): warm the offline rebuild-index route's prefetch cache
Problem: RebuildIndexRequest.prefetch_cache (the off-writer-hold census
pre-parse seam from PR #3168) had exactly one caller that ever populated it
-- the daemon's own bulk-rebuild loop (daemon/bulk_rebuild.py). The offline
`polylogue ops maintenance rebuild-index` CLI, and the daemon's own
POST /api/maintenance/rebuild-index HTTP route, both construct a
RebuildIndexRequest without ever threading a cache, so census always paid
the full unwarmed parse cost on those routes. The operator's real 4h22m CLI
rebuild measured spill_load (serial re-deserialization inline on the writer
thread) at 2,830s -- 70% of the run's parse_s.

Solution: relocated the parse-stage warmer (formerly
polylogue.daemon.parse_prefetch.DaemonParseStage, consulted by exactly one
caller) to substrate as polylogue.sources.census_parse_stage.CensusParseStage
-- it had no daemon-specific behavior, only a daemon-specific import path.
polylogue.daemon.parse_prefetch now re-exports the same objects unchanged so
every existing daemon call site and test keeps working byte-identically
(verified: tests/unit/daemon/test_parse_prefetch.py and 5 other daemon/
sources/storage test files, 110 tests, all green).

maintenance/rebuild_index.py's _rebuild_index_from_source_owned now calls a
new _warm_offline_prefetch_cache helper whenever a caller leaves
request.prefetch_cache at its None default, warming exactly this pass's own
selected_raw_ids in a bounded thread pool before replay. This fixes BOTH the
offline CLI route and the daemon's own HTTP rebuild-index route in one place
(both call the same engine function), rather than needing two separate
call-site patches -- and composes with PR #3468's prepare_session_rows
threading without touching it. daemon/bulk_rebuild.py's own external
off-writer-hold warm is untouched (its caller-supplied cache is never
overridden).

New test (test_rebuild_index_from_source_sync_warms_prefetch_cache_when_
caller_omits_one) drives the real rebuild_index_from_source_sync entry
point with a request that omits prefetch_cache and asserts
CensusParseStage.warm_raw_ids was actually reached with this pass's raw
ids -- verified to fail when the fix is reverted (temporarily disabled the
internal warm call, confirmed the new assertion fails with "never warmed a
prefetch cache", then restored it).

Measurement: a synthetic 400-raw/60KB corpus (tests.infra.
revision_backfill_benchmark) shows census stage time collapsing from ~0.8s
to ~0.1-0.15s with warming enabled -- but spill_load stayed 0.0s in both
runs, because this corpus is far too small to overflow the ~256MiB-2GiB
decoded-session hot cache that gates whether spill_load pays sqlite
pickle.loads at all. This harness measurably under-represents the real
whale-corpus path the operator's receipt came from (per the task brief:
~10% untimed synthetic work vs ~33% real) -- the census-time collapse is
real and directly measured; the spill_load collapse the real 87K-raw run
would show is not independently reproduced here and should not be claimed
as measured.

Follow-up filed: polylogue-mznm (the daemon's own whale-scale raw-
materialization pass has the identical prefetch_cache=None gap against its
sibling trickle pass -- but needs a NEW warm step, not just threading an
existing variable, so scoped out of this PR); polylogue-q1at (proposal to
make this defect class -- a capability adopted by one caller of a shared
engine while siblings keep the disabled default -- detectable by tooling).

Verification: devtools verify --quick (ruff format/check, mypy --strict,
render all --check, topology, layering, all ok); devtools test across the
12 affected daemon/sources/storage/maintenance/cli test files (110 passed);
anti-vacuity check on the new test (fails when the fix is disabled).
